fix bug that removes all file when no api specified

+++ sendFunc.py
@@ -33,6 +33,8 @@ async def luboComment(userConfig, streamInfo):
                 with open("exception.txt", "a", encoding="utf-8") as log:
                     log.write("{} not found for {}\n".format(filename, latest["title"]))
             streamInfo["last_comment_aids"][repo] = latest["aid"]
+    if lesser_file == "9": # in case api returns unexpected response
+        return
     for file in os.listdir("danmu/{}".format(userConfig["ROOM_NAME"])):
         if file <= lesser_file:
             os.remove("danmu/{}/{}".format(userConfig["ROOM_NAME"], file))
